+++ docs/history.rst
@@ -1,6 +1,13 @@
 History
 =======
 
+0.1.12 -- unreleased
+
+- enhancement: enable tab-completion of recently seen words in output
+- bugfix: issuing multiple travel commands should allow canceling
+- bugfix: backspace and delete character codes corrected for input and output of ATASCII/PETSCII
+
+
 0.1.11 -- 2026-08-07
 --------------------